Move DAG to pipeline, where it is only used
