+++ src/main/scala/org/codefeedr/pipeline/Pipeline.scala
@@ -20,7 +20,7 @@ package org.codefeedr.pipeline
 
 import org.apache.flink.api.java.utils.ParameterTool
 import org.apache.flink.streaming.api.scala.{DataStream, StreamExecutionEnvironment}
-import org.codefeedr.{DirectedAcyclicGraph, Properties}
+import org.codefeedr.Properties
 import org.codefeedr.keymanager.KeyManager
 import org.codefeedr.pipeline.buffer.BufferType.BufferType
 import org.codefeedr.pipeline.RuntimeType.RuntimeType
