+++ src/main/scala/org/codefeedr/pipeline/PipelineObject.scala
@@ -22,7 +22,6 @@ import com.sksamuel.avro4s.FromRecord
 import org.apache.flink.api.java.operators.DataSink
 import org.apache.flink.streaming.api.functions.sink.SinkFunction
 import org.apache.flink.streaming.api.scala.DataStream
-import org.codefeedr.DirectedAcyclicGraph
 import org.codefeedr.pipeline.buffer.BufferFactory
 
 import scala.reflect.{ClassTag, Manifest}
