+++ src/main/scala/org/codefeedr/pipeline/PipelineObjectN.scala
@@ -20,7 +20,6 @@ package org.codefeedr.pipeline
 
 import com.sksamuel.avro4s.FromRecord
 import org.apache.flink.streaming.api.scala.DataStream
-import org.codefeedr.DirectedAcyclicGraph
 
 import scala.reflect.{ClassTag, Manifest}
 import scala.reflect.runtime.universe._
